backend/requirements.txt
daphne==4.0.0
Django==4.1.7
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
Markdown==3.4.1
pytz==2022.7.1
psycopg2-binary==2.9.5
Pillow==9.4.0
agent==0.1.2
